Updated home type model. Moved to house database. Also, made it a foreign key instead of char field for the homes
